Split unix/windows stuff
