+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Runtime/InteropServices/NativeLibrary.NativeAot.cs
@@ -1,10 +1,8 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics;
 using System.IO;
 using System.Reflection;
-using System.Text;
 
 using LibraryNameVariation = System.Runtime.Loader.LibraryNameVariation;
 
@@ -116,141 +114,13 @@ namespace System.Runtime.InteropServices
             return ret;
         }
 
-        private static IntPtr LoadLibraryHelper(string libraryName, int flags, ref LoadLibErrorTracker errorTracker)
-        {
-#if TARGET_WINDOWS
-            IntPtr ret = Interop.Kernel32.LoadLibraryEx(libraryName, IntPtr.Zero, flags);
-            if (ret != IntPtr.Zero)
-            {
-                return ret;
-            }
-
-            int lastError = Marshal.GetLastWin32Error();
-            if (lastError != Interop.Errors.ERROR_INVALID_PARAMETER)
-            {
-                errorTracker.TrackErrorCode(lastError);
-            }
-
-            return ret;
-#else
-            // TODO: FileDosToUnixPathA
-            IntPtr ret = Interop.Sys.LoadLibrary(libraryName);
-            if (ret == IntPtr.Zero)
-            {
-                string? message = Marshal.PtrToStringAnsi(Interop.Sys.GetLoadLibraryError());
-                errorTracker.TrackErrorMessage(message);
-            }
-
-            return ret;
-#endif
-        }
-
-        private static void FreeLib(IntPtr handle)
-        {
-            Debug.Assert(handle != IntPtr.Zero);
-
-#if !TARGET_UNIX
-            bool result = Interop.Kernel32.FreeLibrary(handle);
-            if (!result)
-                throw new InvalidOperationException();
-#else
-            Interop.Sys.FreeLibrary(handle);
-#endif
-        }
-
         private static unsafe IntPtr GetSymbol(IntPtr handle, string symbolName, bool throwOnError)
         {
-            IntPtr ret;
-#if !TARGET_UNIX
-            var symbolBytes = new byte[Encoding.UTF8.GetByteCount(symbolName) + 1];
-            Encoding.UTF8.GetBytes(symbolName, symbolBytes);
-            fixed (byte* pSymbolBytes = symbolBytes)
-            {
-                ret = Interop.Kernel32.GetProcAddress(handle, pSymbolBytes);
-            }
-#else
-            ret = Interop.Sys.GetProcAddress(handle, symbolName);
-#endif
+            IntPtr ret = GetSymbolOrNull(handle, symbolName);
             if (throwOnError && ret == IntPtr.Zero)
                 throw new EntryPointNotFoundException(SR.Format(SR.Arg_EntryPointNotFoundExceptionParameterizedNoLibrary, symbolName));
 
             return ret;
-        }
-
-        // Preserving good error info from DllImport-driven LoadLibrary is tricky because we keep loading from different places
-        // if earlier loads fail and those later loads obliterate error codes.
-        //
-        // This tracker object will keep track of the error code in accordance to priority:
-        //
-        //   low-priority:      unknown error code (should never happen)
-        //   medium-priority:   dll not found
-        //   high-priority:     dll found but error during loading
-        //
-        // We will overwrite the previous load's error code only if the new error code is higher priority.
-        internal struct LoadLibErrorTracker
-        {
-#if TARGET_WINDOWS
-            private int _errorCode;
-            private int _priority;
-
-            private const int PriorityNotFound = 10;
-            private const int PriorityAccessDenied = 20;
-            private const int PriorityCouldNotLoad = 99999;
-
-            public void Throw(string libraryName)
-            {
-                if (_errorCode == Interop.Errors.ERROR_BAD_EXE_FORMAT)
-                {
-                    throw new BadImageFormatException();
-                }
-
-                string message = Interop.Kernel32.GetMessage(_errorCode);
-                throw new DllNotFoundException(SR.Format(SR.DllNotFound_Windows, libraryName, message));
-            }
-
-            public void TrackErrorCode(int errorCode)
-            {
-                int priority = errorCode switch
-                {
-                    Interop.Errors.ERROR_FILE_NOT_FOUND or
-                    Interop.Errors.ERROR_PATH_NOT_FOUND or
-                    Interop.Errors.ERROR_MOD_NOT_FOUND or
-                    Interop.Errors.ERROR_DLL_NOT_FOUND => PriorityNotFound,
-
-                    // If we can't access a location, we can't know if the dll's there or if it's good.
-                    // Still, this is probably more unusual (and thus of more interest) than a dll-not-found
-                    // so give it an intermediate priority.
-                    Interop.Errors.ERROR_ACCESS_DENIED => PriorityAccessDenied,
-
-                    // Assume all others are "dll found but couldn't load."
-                    _ => PriorityCouldNotLoad,
-                };
-
-                if (priority > _priority)
-                {
-                    _errorCode = errorCode;
-                    _priority = priority;
-                }
-            }
-#else
-            // On Unix systems we don't have detailed programatic information on why load failed
-            // so there's no priorities.
-            private string? _errorMessage;
-
-            public void Throw(string libraryName)
-            {
-#if TARGET_OSX
-                throw new DllNotFoundException(SR.Format(SR.DllNotFound_Mac, libraryName, _errorMessage));
-#else
-                throw new DllNotFoundException(SR.Format(SR.DllNotFound_Linux, libraryName, _errorMessage));
-#endif
-            }
-
-            public void TrackErrorMessage(string? message)
-            {
-                _errorMessage = message;
-            }
-#endif
         }
     }
 }
